% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen_temperatura.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\alias{Calcula}
\alias{la}
\alias{media}
\alias{y}
\alias{el}
\alias{desvío}
\alias{estándar}
\alias{de}
\alias{temperatura}
\alias{abrigo}
\alias{a}
\alias{150}
\alias{cm}
\alias{para}
\alias{un}
\alias{conjunto}
\alias{estaciones.}
\title{Tabla resumen de temperatura por estación}
\usage{
tabla_resumen_temperatura(datos, estaciones)
}
\arguments{
\item{datos}{data.frame con los datos (debe tener columnas \code{id}
y \code{temperatura_abrigo_150cm}).}

\item{estaciones}{vector de ids de estaciones a incluir.}
}
\value{
data.frame con columnas: id, promedio, desviacion.
}
\description{
Tabla resumen de temperatura por estación
}
\examples{
data("datos_ejemplo")
tabla_resumen_temperatura(
  datos_ejemplo,
  estaciones = c("NHO472", "NHO437")
)

}
